add rating dialog to the media details screen

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/reusableinteractionlistener/rate/RateInteractionListener.kt
@@ -3,6 +3,6 @@ package com.berlin.aflami.viewmodel.reusableinteractionlistener.rate
 interface RateInteractionListener {
     fun onRateIconClicked(id: Long)
     fun onSelectRateClicked(rate: Float)
-    fun onSubmitRateClicked(rate: Float)
+    fun onSubmitRateClicked(rate: Int)
     fun onCancelRatingClicked()
 }
